Merge


Former-commit-id: 43266fd4b3f23af9391b02cc2798a822394afa20

+++ src/game_data/states/State.java
@@ -3,6 +3,6 @@ package game_data.states;
 public interface State {
 	
 	public State copy();
-	public void updateState(int pain);
+	public void updateState(double pain);
 
 }
